pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-web-scraper"
version = "1.0.0"
description = "A configurable, production-ready web scraper with YAML-based configuration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Vladyslav Zdvizhkov"}]
keywords = ["web-scraping", "beautifulsoup", "httpx", "async", "configurable"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "click>=8.1",
    "pydantic>=2.5",
    "pandas>=2.1",
    "pyyaml>=6.0",
    "tenacity>=8.2",
    "openpyxl>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "respx>=0.21",
]

[project.scripts]
smart-scraper = "scraper.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
